hal/ipa/NeonStatsEncoder: integer sharpness accumulator
The Sobel squared-gradient hot loop was widening each 8-tap uint16x8
to two int32x4 halves (16 vmovl_u16 instructions), doing Sobel in
s32, converting the result to float (4 vcvtq_f32_s32) and
accumulating with vmlaq_f32. |gx|, |gy| fit in s16 comfortably
(4·1023 = 4092), so keep the Sobel math in s16 throughout and use
vmlal_s16 to square-and-widen into an int32 accumulator in one op.

Drops per chunk (8 G pixels):
 - 16 vmovl_u16 widens → 8 vreinterpret (free).
 - ~24 vaddq_s32 / vshlq_n_s32 in s32 → ~14 vaddq_s16 / vshlq_n_s16.
 - 4 vcvtq_f32_s32 + 4 vmlaq_f32 → 4 vmlal_s16.

Per-patch Tenengrad max is ~1.3e11 on high-frequency scenes — still
inside uint64. Keep Partial::sharpSum as u64, fold the int32 NEON
accumulator through u32 hsum at the end of each patch cell; finalize
casts u64 → float for the IpaStats field so BasicIpa's consumer
interface does not change. At the u64 magnitude the float cast
loses ~5 decimal digits of precision, which is well below the
dynamic range any AF peak-detect comparator cares about.

Scalar fallback in scalarSobelStep moves off float as well — gx²+gy²
fits in int32, only the per-patch sum needs u64.

+++ hal/ipa/NeonStatsEncoder.h
@@ -41,7 +41,12 @@ public:
     struct Partial {
         uint64_t sumCh   [IpaStats::PATCH_Y][IpaStats::PATCH_X][3];
         uint32_t cntCh   [IpaStats::PATCH_Y][IpaStats::PATCH_X][3];
-        float    sharpSum[IpaStats::PATCH_Y][IpaStats::PATCH_X];
+        /* Tenengrad sum of squared Sobel gradients per patch. Kept as
+         * uint64_t so the inner NEON kernel can accumulate in s32 and
+         * fold into u64 per patch cell without a float conversion on
+         * the hot path — see the vmlal_s16 accumulator in computeRange.
+         * finalize() casts down to the IpaStats float field. */
+        uint64_t sharpSum[IpaStats::PATCH_Y][IpaStats::PATCH_X];
         uint32_t lumaHist[IpaStats::HIST_BINS];
     };
